backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.10.0
yt-dlp==2024.12.23
APScheduler==3.10.4
python-multipart==0.0.18
